+++ packages/io/amf-deserializer/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.1](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/amf-deserializer@2.2.0...@jscad/amf-deserializer@2.2.1) (2021-10-04)
+
+**Note:** Version bump only for package @jscad/amf-deserializer
+
+
+
+
+
+# [2.2.0](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/amf-deserializer@2.1.3...@jscad/amf-deserializer@2.2.0) (2021-09-27)
+
+
+### Features
+
+* **amf-deserializer:** reworked to use saxes library ([8e89118](https://github.com/jscad/OpenJSCAD.org/commit/8e89118337499575da5613d8aaf86b78241ccfe9))
+
+
+
+
+
 ## [2.1.3](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/amf-deserializer@2.1.2...@jscad/amf-deserializer@2.1.3) (2021-09-09)